chore: remove extra _gcs and _types files
