+++ pathy/_tests/__init__.py
@@ -1,0 +1,7 @@
+has_gcs: bool
+try:
+    from ..gcs import BucketClientGCS
+
+    has_gcs = bool(BucketClientGCS)
+except ImportError:
+    has_gcs = False
