+++ pathy/_tests/conftest.py
@@ -6,9 +6,9 @@ from pathlib import Path
 from typing import Any, Generator, Optional
 
 import pytest
-
 from pathy import Pathy, set_client_params, use_fs, use_fs_cache
-from pathy.gcs import has_gcs
+
+from . import has_gcs
 
 has_credentials = "GCS_CREDENTIALS" in os.environ
