+++ pathy/_tests/test_gcs.py
@@ -3,7 +3,8 @@ from unittest.mock import patch
 import pytest
 
 from pathy import set_client_params
-from pathy.gcs import has_gcs
+
+from . import has_gcs
 
 
 def raise_default_creds_error() -> None:
