+++ pathy/cli.py
@@ -142,3 +142,5 @@ def ls(
 
 if __name__ == "__main__":
     app()
+
+__all__ = ()
